+++ backend/go.mod
@@ -4,6 +4,7 @@ go 1.23.1
 
 require (
 	cloud.google.com/go/firestore v1.18.0
+	github.com/coder/websocket v1.8.13
 	google.golang.org/api v0.241.0
 )
 
@@ -13,7 +14,6 @@ require (
 	cloud.google.com/go/auth/oauth2adapt v0.2.8 // indirect
 	cloud.google.com/go/compute/metadata v0.7.0 // indirect
 	cloud.google.com/go/longrunning v0.6.7 // indirect
-	github.com/coder/websocket v1.8.13 // indirect
 	github.com/felixge/httpsnoop v1.0.4 // indirect
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
